# project/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Custom Bikes",
        "contact": {
            "name": "Developer Contact",
            "url": "https://www.example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/transitsNotRelated": {
            "get": {
                "tags": [
                    "Transits"
                ],
                "summary": "List Transits which are not related",
                "responses": {
                    "200": {
                        "description": "Success: List all Transits which are not related",
                        "schema": {
                            "$ref": "#/definitions/Transit"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/transits": {
            "get": {
                "tags": [
                    "Transits"
                ],
                "summary": "List Transits",
                "responses": {
                    "200": {
                        "description": "Success: List all Transits",
                        "schema": {
                            "$ref": "#/definitions/Transit"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Transits"
                ],
                "summary": "Create Transit",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Transit"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success: A Newly Created Transit",
                        "schema": {
                            "$ref": "#/definitions/Transit"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            }
        },
        "/api/transits/{id}": {
            "get": {
                "tags": [
                    "Transits"
                ],
                "summary": "Get Transit by Id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Display the specified Transit by id.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Transit",
                        "schema": {
                            "$ref": "#/definitions/Transit"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            },
            "put": {
                "tags": [
                    "Transits"
                ],
                "summary": "Update Transit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Update the specified Transit by id.",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Transit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Transit updated",
                        "schema": {
                            "$ref": "#/definitions/Transit"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transits"
                ],
                "summary": "Delete Transit",
                "description": "Delete the specified Transit by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Transit id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    },
                    "204": {
                        "description": "Success: successful deleted"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Transit": {
            "required": [
                "navire",
                "fournisseur",
                "marchandise",
                "dateArrivee",
                "franchise"
            ],
            "properties": {
                "navire_id": {
                    "description": "designation d\u00e9crivant le navire du Transit",
                    "type": "string",
                    "example": "1"
                },
                "contenance": {
                    "description": "designation d\u00e9crivant la contenance du Transit",
                    "type": "string",
                    "example": "24"
                },
                "societe_id": {
                    "description": "designation d\u00e9crivant le fournisseur du Transit",
                    "type": "string",
                    "example": "1"
                },
                "marchandise_id": {
                    "description": "designation d\u00e9crivant la marchandise du Transit",
                    "type": "string",
                    "example": "1"
                },
                "dateArrivee": {
                    "description": "designation d\u00e9crivant la date d'arriv\u00e9e du Transit",
                    "type": "string",
                    "example": "2011/10/21"
                },
                "franchise": {
                    "description": "designation d\u00e9crivant la franchise du Transit",
                    "type": "string",
                    "example": "27"
                }
            }
        },
...
    }
}